Update: refactored and connected Dario's background and animations. Als fixed tutorial push fade in and out, also made end scene trigger start screen and updated the temp texture, Also also updated a lot of sorting orders, to make sure everything renders correctly. also also also, made hold zoom in stop when you let go early

+++ Assets/Scripts/Organism/Bucket.cs
@@ -21,7 +21,7 @@ public class Bucket : MonoBehaviour
     private OrganismAnim organismToFollow;
     private int savedAmount;
     public int CurrentState => animateAlongSpline.CurrentState;
-    public Action<int> OnEmptyBucket;
+    public Action<float> OnEmptyBucket;
 
     private void OnValidate()
     {
